+++ src/main/java/org/tech_solutions/application/accounts/dto/AccountDataDTO.java
@@ -3,12 +3,11 @@ package org.tech_solutions.application.accounts.dto;
 import org.tech_solutions.application.accounts.enums.Type;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 public record AccountDataDTO(
-        BigInteger id,
-        BigInteger userId,
+        Long id,
+        Long userId,
         String name,
         Type type,
         BigDecimal balance,
@@ -16,4 +15,5 @@ public record AccountDataDTO(
         LocalDateTime updatedAt
 ) {
 }
+
