+++ src/main/java/org/tech_solutions/application/accounts/dto/AccountRequestDTO.java
@@ -6,11 +6,10 @@ import jakarta.validation.constraints.NotNull;
 import org.tech_solutions.application.accounts.enums.Type;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 
 public record AccountRequestDTO(
         @NotNull(message = "'userId' nao pode ser nulo")
-        BigInteger userId,
+        Long userId,
         @NotBlank(message = "'name' nao pode ser vazio")
         String name,
         @NotNull(message = "'type' nao pode ser nulo")
@@ -20,4 +19,5 @@ public record AccountRequestDTO(
         BigDecimal balance
 ) {
 }
+
