+++ src/main/java/org/tech_solutions/application/accounts/repository/AccountRepository.java
@@ -4,11 +4,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.tech_solutions.application.accounts.model.Account;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface AccountRepository extends JpaRepository<Account, BigInteger> {
-    List<Account> findByUserId(BigInteger userId);
+public interface AccountRepository extends JpaRepository<Account, Long> {
+    List<Account> findByUserId(Long userId);
 }
+
