+++ src/main/java/org/tech_solutions/application/aiinsights/dto/AiInsightDataDTO.java
@@ -2,15 +2,15 @@ package org.tech_solutions.application.aiinsights.dto;
 
 import org.tech_solutions.application.aiinsights.enums.InsightType;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 public record AiInsightDataDTO(
-        BigInteger id,
-        BigInteger userId,
+        Long id,
+        Long userId,
         InsightType insightType,
         String content,
         LocalDateTime generatedAt
 ) {
 }
+
