+++ src/main/java/org/tech_solutions/application/aiinsights/model/AiInsight.java
@@ -4,7 +4,6 @@ import jakarta.persistence.*;
 import org.tech_solutions.application.aiinsights.enums.InsightType;
 import org.tech_solutions.application.user.model.User;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 @Entity
@@ -13,7 +12,7 @@ public class AiInsight {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigInteger id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
@@ -32,11 +31,11 @@ public class AiInsight {
     public AiInsight() {
     }
 
-    public BigInteger getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -72,4 +71,5 @@ public class AiInsight {
         this.generatedAt = generatedAt;
     }
 }
+
