+++ src/main/java/org/tech_solutions/application/aiinsights/repository/AiInsightRepository.java
@@ -4,11 +4,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.tech_solutions.application.aiinsights.model.AiInsight;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface AiInsightRepository extends JpaRepository<AiInsight, BigInteger> {
-    List<AiInsight> findByUserId(BigInteger userId);
+public interface AiInsightRepository extends JpaRepository<AiInsight, Long> {
+    List<AiInsight> findByUserId(Long userId);
 }
+
