+++ src/main/java/org/tech_solutions/application/categories/dto/CategoryDataDTO.java
@@ -2,15 +2,15 @@ package org.tech_solutions.application.categories.dto;
 
 import org.tech_solutions.application.categories.enums.CategoryType;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 public record CategoryDataDTO(
-        BigInteger id,
-        BigInteger userId,
+        Long id,
+        Long userId,
         String name,
         CategoryType type,
         LocalDateTime createdAt
 ) {
 }
+
