+++ src/main/java/org/tech_solutions/application/categories/dto/CategoryRequestDTO.java
@@ -4,15 +4,15 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import org.tech_solutions.application.categories.enums.CategoryType;
 
-import java.math.BigInteger;
 
 public record CategoryRequestDTO(
         @NotNull(message = "'userId' nao pode ser nulo")
-        BigInteger userId,
+        Long userId,
         @NotBlank(message = "'name' nao pode ser vazio")
         String name,
         @NotNull(message = "'type' nao pode ser nulo")
         CategoryType type
 ) {
 }
+
