+++ src/main/java/org/tech_solutions/application/importedtransactions/dto/ImportedTransactionDataDTO.java
@@ -1,16 +1,16 @@
 package org.tech_solutions.application.importedtransactions.dto;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.time.LocalDate;
 
 public record ImportedTransactionDataDTO(
-        BigInteger id,
-        BigInteger importId,
+        Long id,
+        Long importId,
         String rawDescription,
         BigDecimal rawAmount,
         LocalDate rawDate,
         Boolean processed
 ) {
 }
+
