+++ src/main/java/org/tech_solutions/application/importedtransactions/dto/ImportedTransactionRequestDTO.java
@@ -3,16 +3,16 @@ package org.tech_solutions.application.importedtransactions.dto;
 import jakarta.validation.constraints.NotNull;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.time.LocalDate;
 
 public record ImportedTransactionRequestDTO(
         @NotNull(message = "'importId' nao pode ser nulo")
-        BigInteger importId,
+        Long importId,
         String rawDescription,
         BigDecimal rawAmount,
         LocalDate rawDate,
         Boolean processed
 ) {
 }
+
