+++ src/main/java/org/tech_solutions/application/importedtransactions/repository/ImportedTransactionRepository.java
@@ -4,11 +4,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.tech_solutions.application.importedtransactions.model.ImportedTransaction;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface ImportedTransactionRepository extends JpaRepository<ImportedTransaction, BigInteger> {
-    List<ImportedTransaction> findByImportFileId(BigInteger importId);
+public interface ImportedTransactionRepository extends JpaRepository<ImportedTransaction, Long> {
+    List<ImportedTransaction> findByImportFileId(Long importId);
 }
+
