+++ src/main/java/org/tech_solutions/application/imports/dto/ImportDataDTO.java
@@ -2,15 +2,15 @@ package org.tech_solutions.application.imports.dto;
 
 import org.tech_solutions.application.imports.enums.ImportStatus;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 public record ImportDataDTO(
-        BigInteger id,
-        BigInteger userId,
+        Long id,
+        Long userId,
         String fileName,
         LocalDateTime importedAt,
         ImportStatus status
 ) {
 }
+
