+++ src/main/java/org/tech_solutions/application/imports/repository/ImportFileRepository.java
@@ -4,11 +4,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.tech_solutions.application.imports.model.ImportFile;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface ImportFileRepository extends JpaRepository<ImportFile, BigInteger> {
-    List<ImportFile> findByUserId(BigInteger userId);
+public interface ImportFileRepository extends JpaRepository<ImportFile, Long> {
+    List<ImportFile> findByUserId(Long userId);
 }
+
