+++ src/main/java/org/tech_solutions/application/transactions/dto/TransactionDataDTO.java
@@ -3,15 +3,14 @@ package org.tech_solutions.application.transactions.dto;
 import org.tech_solutions.application.transactions.enums.TransactionType;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 public record TransactionDataDTO(
-        BigInteger id,
-        BigInteger userId,
-        BigInteger accountId,
-        BigInteger categoryId,
+        Long id,
+        Long userId,
+        Long accountId,
+        Long categoryId,
         String transactionDescription,
         BigDecimal amount,
         LocalDate transactionDate,
@@ -19,4 +18,5 @@ public record TransactionDataDTO(
         LocalDateTime createdAt
 ) {
 }
+
